backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pydantic[email]==2.10.6
pydantic-settings==2.7.1
cryptography==44.0.1
redis==5.2.1
